+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Redirect to `/shipping` after profile completion.
+- Click on checkboxes label (save info and newsletter opt-in) to toggle the checkbox state.
 
 ## [0.1.0] - 2020-03-30
 ### Added